Create Search Component(autocomplete,filtering)
