+++ app/[lang]/page.tsx
@@ -14,7 +14,6 @@ export default async function Home() {
       <AboutFoundation />
       <OurMission />
       <OurGoals />
-      <Compositions />
       <LiatoshynskyOffice />
       <WhatWeDo />
       <FoundationFounders />
